requirements.txt
certifi==2020.6.20
chardet==3.0.4
docopt==0.6.2
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1
openpyxl==2.4.11
prettytable==1.0.1
protobuf==3.13.0
records==0.5.3
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
tablib==2.0.0
urllib3==1.25.10
wcwidth==0.2.5
